Split Ollama entity
